fix: jaccard 알고리즘 보완

+++ src/main/java/com/openpromt/coffeee/swf2023/openpromtserver/copyright/dto/RegisterCopyrightResponse.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class RegisterCopyrightResponse {
     private String copyright_name;
-    private Integer similarity;
+    private Double similarity;
     private Boolean validate;
 }
